KTLN-204: Move PrivateSetterExamples to core-kotlin-7
* core-kotlin-3 has already 10 articles
