+++ core-kotlin-modules/core-kotlin-7/src/main/kotlin/com/baeldung/privatesetter/BankAccount.kt
