+++ core-kotlin-modules/core-kotlin-7/src/main/kotlin/com/baeldung/privatesetter/PrivateSetterExamples.kt
@@ -1,6 +1,10 @@
 package com.baeldung.privatesetter
 
 
+fun main() {
+    println("Private Setter Examples")
+}
+
 object PrivateSetterExamples {
 
 //    fun myClassExample() {
